cleaned up some more
